+++ crates/agent-tunnel-proto/src/version.rs
@@ -1,8 +1,8 @@
 /// Current protocol version.
-pub const CURRENT_PROTOCOL_VERSION: u16 = 2;
+pub const CURRENT_PROTOCOL_VERSION: u16 = 1;
 
 /// Minimum protocol version that is still accepted.
-pub const MIN_SUPPORTED_VERSION: u16 = 2;
+pub const MIN_SUPPORTED_VERSION: u16 = 1;
 
 /// Validate that a received protocol version is within the supported range.
 pub fn validate_protocol_version(version: u16) -> Result<(), crate::error::ProtoError> {
