+++ devolutions-agent/Cargo.toml
@@ -14,6 +14,7 @@ workspace = true
 [dependencies]
 agent-tunnel-proto = { path = "../crates/agent-tunnel-proto" }
 anyhow = "1"
+backoff = "0.4"
 async-trait = "0.1"
 bincode = "1.3"
 base64 = "0.22"
